requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0
djangorestframework==3.12.4
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
